Split default page and search page to keep the `/hall-of-fame` static
