+++ apps/bestofjs-nextjs/src/app/hall-of-fame/page.tsx
@@ -2,23 +2,16 @@ import { Metadata } from "next";
 
 import { db } from "@repo/db";
 import { findHallOfFameMembers } from "@repo/db/hall-of-fame";
-import { ExternalLink, PageHeading } from "@/components/core/typography";
-import {
-  APP_CANONICAL_URL,
-  APP_DISPLAY_NAME,
-  APP_REPO_URL,
-} from "@/config/site";
+import { APP_CANONICAL_URL, APP_DISPLAY_NAME } from "@/config/site";
 import { HallOfFamePaginatedList } from "./hall-of-fame-paginated-list";
-import { HallOfFameSearchStateParser } from "./hall-of-fame-search-state";
 import Loading from "./loading";
-import { HallOfFameSearchBar } from "./search-bar";
+import { HallOfFameSearchStateParser } from "./search/hall-of-fame-search-state";
+import { HallOfFameSearchBar } from "./search/search-bar";
 
 const forceLoadingState = false; // set to true when debugging the loading state
 
 const description =
   "Some of the greatest developers, authors and speakers of the JavaScript community. Meet Evan, Dan, Sindre, TJ and friends!";
-
-const searchStateParser = new HallOfFameSearchStateParser();
 
 export async function generateMetadata(): Promise<Metadata> {
   return {
@@ -33,42 +26,21 @@ export async function generateMetadata(): Promise<Metadata> {
   };
 }
 
-type PageProps = {
-  searchParams: { query?: string };
-};
+const searchStateParser = new HallOfFameSearchStateParser();
 
-export default async function HallOfFamePage({ searchParams }: PageProps) {
+export default async function HallOfFamePage() {
   if (forceLoadingState) return <Loading />;
 
-  const { searchState, buildPageURL } = searchStateParser.parse(searchParams);
+  const { searchState, buildPageURL } = searchStateParser.parse({});
 
   const { members, total } = await findHallOfFameMembers({
     db,
-    page: searchState.page,
-    limit: searchState.limit,
-    searchQuery: searchState.query,
+    page: 1,
+    limit: 50,
   });
 
   return (
-    <div className="space-y-6">
-      <PageHeading
-        title={<>JavaScript Hall of Fame</>}
-        subtitle={
-          <>
-            Here are some of the greatest developers, authors and speakers of
-            the JavaScript community.
-            <br />
-            They are sorted by number of followers,{" "}
-            <ExternalLink
-              url={APP_REPO_URL}
-              className="color-primary underline hover:no-underline"
-            >
-              contact us
-            </ExternalLink>{" "}
-            to add more members!
-          </>
-        }
-      />
+    <>
       <HallOfFameSearchBar />
       <HallOfFamePaginatedList
         members={members}
@@ -76,6 +48,6 @@ export default async function HallOfFamePage({ searchParams }: PageProps) {
         searchState={searchState}
         buildPageURL={buildPageURL}
       />
-    </div>
+    </>
   );
 }
